[ci skip] Build fix

+++ include/jet/constants.h
@@ -344,15 +344,15 @@ constexpr double invTwoPi<double>() {
 // MARK: Physics
 
 //! Gravity.
-constexpr double kGravityF = -9.8f;
+constexpr float kGravityF = -9.8f;
 constexpr double kGravityD = -9.8;
 
 //! Water density.
-constexpr double kWaterDensityF = 1000.0f;
+constexpr float kWaterDensityF = 1000.0f;
 constexpr double kWaterDensityD = 1000.0;
 
 //! Speed of sound in water at 20 degrees celcius.
-constexpr double kSpeedOfSoundInWaterF = 1482.0f;
+constexpr float kSpeedOfSoundInWaterF = 1482.0f;
 constexpr double kSpeedOfSoundInWaterD = 1482.0;
 
 // MARK: Common enums
